Adding Anime fill missing frames only MODE

+++ inference_dain.py
@@ -30,6 +30,16 @@ def config():
         type=bool,
         help="split the frames when handling 720+ videos",
     )
+    parser.add_argument(
+        "--input_video_fps", "-ifps", type=float, help="Input Video FPS for Missing Frames only Anime MODE"
+    )
+    parser.add_argument(
+        "--anime_fill_frames",
+        "-aff",
+        default=False,
+        type=bool,
+        help="Fill in only the missing frames",
+    )
     return parser.parse_args()
 
 
@@ -37,33 +47,41 @@ if __name__ == "__main__":
     os.chdir(DAIN_PREFIX)
     print(f"Current PyTorch version is {torch.__version__}")
     args = config()
-    assert args.time_step in [
-        0.5,
-        0.25,
-        0.125,
-    ], "Time step must be one of 0.5/0.25/0.125."
+    
+    if not args.anime_fill_frames:
+        assert args.time_step in [
+            0.5,
+            0.25,
+            0.125,
+        ], "Time step must be one of 0.5/0.25/0.125."
 
-    # STAGE 1: video pre-processing
-    # if len(os.listdir(input_data_dir)) > 1:
-    #     raise FileExistsError("You can only process one video at a time..")
-    video_file_link = os.path.join(input_data_dir, args.input_video)
-    frame_num = frames_info(video_file_link)
-    if frame_num < 2:
-        raise FileNotFoundError(
-            "Video File not found, or incompatible."
+        # STAGE 1: video pre-processing
+        # if len(os.listdir(input_data_dir)) > 1:
+        #     raise FileExistsError("You can only process one video at a time..")
+        video_file_link = os.path.join(input_data_dir, args.input_video)
+        frame_num = frames_info(video_file_link)
+        if frame_num < 2:
+            raise FileNotFoundError(
+                "Video File not found, or incompatible."
+            )
+        fps = fps_info(video_file_link)
+        target_fps = float(fps) / args.time_step
+        video_extract(src=video_file_link, dst=input_data_dir, thread=4)
+        os.remove(video_file_link)
+        print(
+            f"\n--------------------SUMMARY--------------------\n"
+            f"Current input video file is {args.input_video},\n"
+            f"{args.input_video}'s fps is {fps},\n"
+            f"{args.input_video} has {frame_num} frames.\n"
+            f"Now we will process this video to {target_fps} fps.\n"
+            f"--------------------NOW END--------------------\n\n"
         )
-    fps = fps_info(video_file_link)
-    target_fps = float(fps) / args.time_step
-    video_extract(src=video_file_link, dst=input_data_dir, thread=4)
-    os.remove(video_file_link)
-    print(
-        f"\n--------------------SUMMARY--------------------\n"
-        f"Current input video file is {args.input_video},\n"
-        f"{args.input_video}'s fps is {fps},\n"
-        f"{args.input_video} has {frame_num} frames.\n"
-        f"Now we will process this video to {target_fps} fps.\n"
-        f"--------------------NOW END--------------------\n\n"
-    )
+    else:
+        print(
+            f"\n--------------------SUMMARY--------------------\n"
+            f"Interpolating Anime missing frames only!,\n"
+            f"--------------------NOW END--------------------\n\n"
+        )
 
     # STAGE 2: Inference
     os.system(
@@ -71,6 +89,7 @@ if __name__ == "__main__":
         f"--src {input_data_dir} "
         f"--dst {output_data_dir} "
         f"--high_resolution {args.high_resolution} "
+        f"--anime_fill_frames {args.anime_fill_frames} "
         f"--time_step {args.time_step} "
     )
 
@@ -80,11 +99,20 @@ if __name__ == "__main__":
     file_order(src=output_data_dir, dst=input_data_dir)
     output_video_file = f"{args.input_video.split('.')[0]}-{target_fps}.{args.input_video.split('.')[1]}"
     
-    video_fusion(
-        src=input_data_dir + "/%10d.png",
-        dst=os.path.join(output_data_dir, output_video_file),
-        fps=target_fps,
-        thread=4,
-    )
+    if args.anime_fill_frames:
+        video_fusion(
+            src=input_data_dir + "/%10d.png",
+            dst=os.path.join(output_data_dir, output_video_file),
+            fps=target_fps,
+            thread=4,
+        )
+    else:
+        video_fusion(
+            src=input_data_dir + "/%10d.png",
+            dst=os.path.join(output_data_dir, output_video_file),
+            fps=args.input_video_fps,
+            thread=4,
+        )
+    
     shutil.move(os.path.join(output_data_dir, output_video_file), "/content/gdrive/My Drive/DAIN/Output/" + output_video_file)
     clean_folder(input_data_dir)
